Authorization

+++ controllers/Blog.js
@@ -1,4 +1,5 @@
 const blogRouter = require('express').Router();
+const jwt = require('jsonwebtoken');
 const User = require('../models/User');
 
 const Blog = require('../models/Blog');
@@ -11,7 +12,7 @@ blogRouter.put('/:id', async (req, res, next) => {
       likes,
       url,
     };
-    const updatedBlog = await Blog.findByIdAndUpdate(req.params.id, updateBlog);
+    const updatedBlog = await Blog.findByIdAndUpdate(req.params.id, updateBlog, { new: true });
     res.status(200).json(updatedBlog);
   } catch (error) {
     next(error);
@@ -19,24 +20,37 @@ blogRouter.put('/:id', async (req, res, next) => {
 });
 
 blogRouter.delete('/:id', async (req, res, next) => {
-  try {
+  const { token } = req;
+
+  const decodenToken = jwt.verify(token, process.env.SECRET);
+  const user = await User.findById(decodenToken.id);
+  if (!(token && decodenToken.id)) {
+    res.status(401).json({ error: 'token missing or invalid' });
+  }
+  const blog = await Blog.findById(req.params.id);
+  if (blog.user.toString() === user.id.toString()) {
     await Blog.findByIdAndRemove(req.params.id);
     res.status(204).end();
-  } catch (error) {
-    next(error);
+  } else {
+    res.status(401).json({ error: 'unathorize' });
   }
 });
 
 blogRouter.get('/', async (req, res) => {
-  const blogs = await Blog.find({}).populate('user', {username: 1, id: 1});
+  const blogs = await Blog.find({}).populate('user', { username: 1, id: 1 });
   res.json(blogs);
 });
 
 blogRouter.post('/', async (req, res) => {
+  const { token } = req;
+  const decodenToken = jwt.verify(token, process.env.SECRET);
+  if (!decodenToken.id) {
+    res.status(401).json({ error: 'token missing or invalid' });
+  }
   const {
     title, author, url, likes,
   } = req.body;
-  const user = await User.findOne();
+  const user = await User.findById(decodenToken.id);
 
   const newBlog = new Blog({
     title,
@@ -50,6 +64,11 @@ blogRouter.post('/', async (req, res) => {
   user.blogs = user.blogs.concat(savedBlog._id);
   await user.save();
   res.status(201).json(savedBlog);
+});
+
+blogRouter.get('/:id', async (req, res) => {
+  const blog = await Blog.findById(req.params.id);
+  res.json(blog);
 });
 
 module.exports = blogRouter;
